Move side panels to own include
